instruments/ad_test/requirements.txt
python-ldap
Flask
requests
uwsgi
